feat(design): shard accordion component
