+++ apps/design-land/src/app/accordion/accordion.component.html
@@ -2,11 +2,6 @@
 
 <h3>Types</h3>
 <p><code>daff-accordion-item</code> - An accordion that typically holds content.</p>
-<p><code>daff-nav-accordion-item</code> — An accordion that is used for navigation whose content is nested daff-nav-accordion-items.</p>
 
 <h3>Content Accordion Item</h3>
 <design-land-example-viewer-container example="basic-accordion"></design-land-example-viewer-container>
-
-<h3>(DEPRECATED) Navigation Accordion Item</h3>
-<p>Use the <code>DaffTreeComponent</code> instead.</p>
-<design-land-example-viewer-container example="nav-accordion"></design-land-example-viewer-container>
