+++ libs/design/accordion/examples/src/examples.ts
@@ -1,0 +1,5 @@
+import { BasicAccordionComponent } from './basic-accordion/basic-accordion.component';
+
+export const ACCORDION_EXAMPLES = [
+  BasicAccordionComponent,
+];
