--- libs/design/accordion/examples/src/nav-accordion/nav-accordion.component.html
@@ -1,23 +1,0 @@
-<daff-accordion>
-	<daff-nav-accordion-item>
-		<h3 daffAccordionItemTitle>Details</h3>
-		<daff-nav-accordion-item>
-			<h3 daffAccordionItemTitle>Details</h3>
-			<daff-nav-accordion-item>
-				<a daffAccordionItemTitle>Details</a>
-			</daff-nav-accordion-item>
-			<daff-nav-accordion-item>
-				<a daffAccordionItemTitle>Details</a>
-			</daff-nav-accordion-item>
-		</daff-nav-accordion-item>
-		<daff-nav-accordion-item>
-			<h3 daffAccordionItemTitle>Details</h3>
-			<daff-nav-accordion-item>
-				<a daffAccordionItemTitle>Details</a>
-			</daff-nav-accordion-item>
-			<daff-nav-accordion-item>
-				<a daffAccordionItemTitle>Details</a>
-			</daff-nav-accordion-item>
-		</daff-nav-accordion-item>
-	</daff-nav-accordion-item>
-</daff-accordion>
