--- libs/design/accordion/examples/src/nav-accordion/nav-accordion.module.ts
@@ -1,19 +1,0 @@
-import { NgModule } from '@angular/core';
-
-import { DaffAccordionModule } from '@daffodil/design';
-
-import { NavAccordionComponent } from './nav-accordion.component';
-
-
-@NgModule({
-  declarations: [
-    NavAccordionComponent,
-  ],
-  exports: [
-    NavAccordionComponent,
-  ],
-  imports: [
-    DaffAccordionModule,
-  ],
-})
-export class NavAccordionModule { }
