+++ libs/design/accordion/src/accordion/accordion-item-content/accordion-item-content.directive.ts
