+++ libs/design/accordion/src/accordion/accordion-item-title/accordion-item-title.directive.ts
