+++ libs/design/accordion/src/accordion/accordion-item/accordion-item.component.html
