+++ libs/design/accordion/src/accordion/accordion/accordion.component.html
