+++ libs/design/accordion/src/accordion/animation/accordion-animation-state.spec.ts
