+++ libs/design/accordion/src/accordion/animation/accordion-animation-state.ts
