+++ libs/design/accordion/src/public_api.ts
@@ -1,0 +1,5 @@
+export { DaffAccordionModule } from './accordion.module';
+export * from './accordion/accordion/accordion.component';
+export * from './accordion/accordion-item/accordion-item.component';
+export * from './accordion/accordion-item-title/accordion-item-title.directive';
+export * from './accordion/accordion-item-content/accordion-item-content.directive';
