--- libs/design/src/molecules/accordion/nav-accordion-item/nav-accordion-item.component.html
@@ -1,12 +1,0 @@
-<div class="daff-nav-accordion-item__header" (click)="toggleActive()">
-  <ng-content select="[daffAccordionItemTitle]"></ng-content>
-  <span [hidden]="_open" daffSuffix *ngIf="_navAccordionItemChild.toArray().length">
-    <fa-icon [icon]="faChevronDown"></fa-icon>
-  </span>
-  <span [hidden]="!_open" daffSuffix *ngIf="_navAccordionItemChild.toArray().length">
-    <fa-icon [icon]="faChevronUp"></fa-icon>
-  </span>
-</div>
-<div [@openAccordion]="_animationState">
-  <ng-content></ng-content>
-</div>
