--- libs/design/src/molecules/accordion/nav-accordion-item/nav-accordion-item.component.scss
@@ -1,38 +1,0 @@
-@use '../../../../scss/interactions';
-
-:host(.daff-nav-accordion-item--level-1) {
-	.daff-nav-accordion-item__header {
-		padding-left: 15px;
-	}
-}
-:host(.daff-nav-accordion-item--level-2) {
-	.daff-nav-accordion-item__header {
-		padding-left: 30px;
-	}
-}
-:host(.daff-nav-accordion-item--level-3) {
-	.daff-nav-accordion-item__header {
-		padding-left: 45px;
-	}
-}
-
-:host {
-	display: block;
-	padding-top: 6px;
-}
-
-.daff-nav-accordion-item {
-
-	&__header {
-		@include interactions.clickable;
-		display: flex;
-		align-items: center;
-		justify-content: space-between;
-		width: 100%;
-		padding-bottom: 6px;
-
-		.daff-suffix {
-			margin: 0 10px;
-		}
-	}
-}
