--- libs/design/src/molecules/accordion/public_api.ts
@@ -1,6 +1,0 @@
-export { DaffAccordionModule } from './accordion.module';
-export * from './accordion/accordion.component';
-export * from './accordion-item/accordion-item.component';
-export * from './accordion-item-title/accordion-item-title.directive';
-export * from './accordion-item-content/accordion-item-content.directive';
-export * from './nav-accordion-item/nav-accordion-item.component';
