+++ libs/design/src/public_api.ts
@@ -16,7 +16,6 @@ export * from './atoms/form/radio/public_api';
 export * from './atoms/form/form-label/public_api';
 
 // Molecules
-export * from './molecules/accordion/public_api';
 export * from './molecules/backdrop/public_api';
 export * from './molecules/button-set/public_api';
 export * from './molecules/image-gallery/public_api';
